Swapped in new collector file
